# packages/api/openapi.yaml
openapi: 3.1.0
info:
  title: Fundraising SaaS API
  version: 1.0.0
  description: REST API for the Fundraising SaaS application (Auth, RBAC, CRM, Financial, Documents, Reviewer Portal)
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: /api/v1
    description: Same-origin
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
      description: HttpOnly access token cookie set on login

    reviewerSessionAuth:
      type: apiKey
      in: cookie
      name: reviewerSessionToken
      description: >
        HttpOnly session cookie set after a reviewer completes token + OTP
        verification via POST /reviewer-portal/verify. Required for all
        subsequent Reviewer Portal calls.
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number (1-indexed)
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Invalid credentials
    ValidationErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Validation failed
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: email
              message:
                type: string
                example: Invalid email address
...
    BareMessageResponse:
      type: object
      properties:
        message:
          type: string
          example: Investor removed
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 134
        totalPages:
          type: integer
          example: 7

    # ── Auth ──────────────────────────────────────────────────────────
    # Registration is a two-step OTP flow: /auth/register/initiate stores a
    # pending registration and emails a code, /auth/register/verify turns it
    # into a real user and signs them in. No user row exists in between.
...
    Priority:
      type: string
      enum: [low, medium, high]
      description: Shared by Task.priority (task urgency) and Pipeline.priority (deal importance).
      example: medium
...
    TaskStatus:
      type: string
      enum: [open, completed]
      example: open
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        startupId:
          type: string
          format: uuid
        pipelineId:
          type: string
          format: uuid
          description: The pipeline deal this task belongs to.
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/Priority"
        dueDate:
          type: string
          format: date-time
          nullable: true
        assigneeId:
          type: string
          format: uuid
          nullable: true
          description: The StartupMember id assigned to this task, if any.
        completedAt:
          type: string
          format: date-time
          nullable: true
          description: Set server-side when status becomes "completed"; cleared on reopen.
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateTaskBody:
      type: object
      required: [pipelineId, title]
      properties:
        pipelineId:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        priority:
          $ref: "#/components/schemas/Priority"
        dueDate:
          type: string
          format: date-time
        assigneeId:
          type: string
          format: uuid
          description: A StartupMember id — must belong to the same startup.
    UpdateTaskBody:
      type: object
      description: At least one field is required.
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          nullable: true
          maxLength: 2000
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/Priority"
        dueDate:
          type: string
          format: date-time
          nullable: true
        assigneeId:
          type: string
          format: uuid
          nullable: true
...
paths:
...
  /startups/{startupId}/tasks:
    parameters:
      - name: startupId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Tasks]
      summary: Add a task to a pipeline deal
      operationId: createTask
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Task"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not an active member, or missing pipeline:create
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: PIPELINE_NOT_FOUND or MEMBER_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
    get:
      tags: [Tasks]
      summary: List tasks in a startup
      operationId: listTasks
      description: >
        Filterable by pipelineId, status, assigneeId, and priority — used both
        for a single deal's task list and cross-deal views.
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: pipelineId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/TaskStatus"
        - name: assigneeId
          in: query
          schema:
            type: string
            format: uuid
        - name: priority
          in: query
          schema:
            $ref: "#/components/schemas/Priority"
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Task"
                  meta:
                    $ref: "#/components/schemas/PaginationMeta"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not an active member, or missing pipeline:read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /startups/{startupId}/tasks/{taskId}:
    parameters:
      - name: startupId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: taskId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Tasks]
      summary: Get a task by ID
      operationId: getTask
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Task"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not an active member, or missing pipeline:read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: TASK_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      tags: [Tasks]
      summary: Update a task
      description: >
        Setting status to "completed" stamps completedAt server-side;
        reverting to "open" clears it.
      operationId: updateTask
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Task"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not an active member, or missing pipeline:update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: TASK_NOT_FOUND or MEMBER_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
    delete:
      tags: [Tasks]
      summary: Delete a task
      operationId: deleteTask
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Task removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BareMessageResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not an active member, or missing pipeline:delete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: TASK_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ════════════════════════════════════════════════════════════════
  #  FUNDRAISING ROUNDS (Financial)
  # ════════════════════════════════════════════════════════════════
...